Remove overviewText
Removed `overviewText = "Theseus Inc.";` from `description.ext`.

The text is only shown in the mission select screen, and since we'll be using the `tac_` prefix in front of our missions it's not really required to mark Theseus missions with the overview text.
Additionally, the overviewText will be used for certain missions (mainly non-contracts and gimmicks) to describe certain requirements such as slotting requirements or whatever, however this only works when adding it through EDEN editor (which adds a overviewText entry to mission.sqm). Removing the entry from description.ext just ensures no conflicts.