Enhance template
Overall changed the template.

Instead of a mission file, template is now split into `compositions` and `Contract/Non-Contract Template` folders. Compositions is the spawn, HC etc. that is down ingame, while the template folders hold all the .sqf files etc. that is needed. This is split into Contract and Non-Contract where both have their proper settings set.
For the compositions, it should now be more straightforward to use ingame. Instead of opening the template mission and copying things, players just need to save the composition to their composition folder and they can then quickly pop it down in a new mission.